# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/camps": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "전체 캠프 목록을 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "캠프 목록 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Camp"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "새로운 코너학습 캠프를 생성한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "새 캠프 생성",
                "parameters": [
                    {
                        "description": "캠프 생성 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.CreateCampRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Camp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/camps/{id}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 캠프 정보를 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "캠프 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "캠프 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Camp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "캠프 이름, 예정 기간, 병목 판정 기준 중 요청에 포함된 필드만 수정한다. 종료된 캠프는 수정할 수 없다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "캠프 정보 및 병목 기준 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "캠프 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "부분 수정할 캠프 설정",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.UpdateCampRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Camp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/camps/{id}/end": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "캠프를 ENDED 상태로 변경한다. 이후 데이터 수정이 불가하다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "캠프 종료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "캠프 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Camp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/camps/{id}/start": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "캠프를 ACTIVE 상태로 변경하고 운영을 시작한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "캠프 시작",
                "parameters": [
                    {
                        "type": "string",
                        "description": "캠프 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Camp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "이미 활성화됨 또는 필수 조건 미충족",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_infrastructure_web.Camp": {
            "type": "object",
            "properties": {
                "bottleneckMinSamples": {
                    "type": "integer",
                    "example": 3
                },
                "bottleneckRatioPct": {
                    "type": "integer",
                    "example": 20
                },
                "endAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string",
                    "example": "2026 여름 코너학습"
                },
                "startAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "PENDING",
                        "ACTIVE",
                        "ENDED"
                    ]
                }
            }
        },
...
        "internal_infrastructure_web.CreateCampRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_infrastructure_web.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "잘못된 요청입니다."
                }
            }
        },
...
        "internal_infrastructure_web.UpdateCampRequest": {
            "type": "object",
            "properties": {
                "bottleneckMinSamples": {
                    "type": "integer"
                },
                "bottleneckRatioPct": {
                    "type": "integer"
                },
                "endAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "startAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}